refactor(router): move basic auth to middleware function
